fix: ctrl 1-4 keyboard shortcuts
Refs: #1373
